Move project files to pars_diary directory
