+++ notify.py
@@ -7,13 +7,17 @@ from itertools import zip_longest
 from aiogram import Bot
 from loguru import logger
 
-from models import User
-from utils import db
-from utils.db import DB_NAME
-from utils.exceptions import DBFileNotFoundError, UnknownError, UserNotFoundError
-from utils.load_env import TOKEN
-from utils.pars import Pars
-from utils.typing import UserId
+from pars_diary.models import User
+from pars_diary.utils import db
+from pars_diary.utils.db import DB_NAME
+from pars_diary.utils.exceptions import (
+    DBFileNotFoundError,
+    UnknownError,
+    UserNotFoundError,
+)
+from pars_diary.utils.load_env import TOKEN
+from pars_diary.utils.pars import Pars
+from pars_diary.utils.typing import UserId
 
 # Константы
 # =========
