+++ pars_diary/__main__.py
@@ -1,0 +1,6 @@
+import asyncio
+
+from pars_diary.bot import main
+
+if __name__ == "__main__":
+    asyncio.run(main())
