+++ pars_diary/handlers/admin.py
@@ -12,8 +12,8 @@ from aiogram.filters import Command
 from aiogram.types import FSInputFile, Message
 from loguru import logger
 
-from utils.db import GRAPH_NAME, GetStat, counter, get_graph
-from utils.load_env import ADMINS_TG
+from pars_diary.utils.db import GRAPH_NAME, GetStat, counter, get_graph
+from pars_diary.utils.load_env import ADMINS_TG
 
 router = Router(name=__name__)
 
@@ -42,15 +42,11 @@ async def new_msg(msg: Message) -> None:
             FSInputFile(GRAPH_NAME),
             (
                 f"<b>Всего пользователей: {stat.users_count}</b>\n\n"
-
                 f"Авторизованных пользователей: {stat.cookie}\n\n"
-
                 f"Уведомления: {stat.notify} / {stat.users_count}"
                 f" ({stat.notify / stat.users_count * 100}%)\n"
-
                 f"Умные уведомления: {stat.smart_notify} / {stat.users_count}"
                 f" ({stat.smart_notify / stat.users_count * 100}%)\n\n"
-
                 f"Использований комманды /about: {stat.command_about}\n"
                 f"Использований комманды /admin: {stat.command_admin}\n"
                 f"Использований комманды /birthdays: {stat.command_birthdays}\n"
@@ -64,7 +60,6 @@ async def new_msg(msg: Message) -> None:
                 f"Использований комманды /new: {stat.command_new}\n"
                 f"Использований комманды /notify: {stat.command_notify_settings}\n"
                 f"Использований комманды /start: {stat.command_start}\n\n"
-
                 "<b>Источники прихода аудитории (рефералы)"
                 " (в порядке уменьшения выгоды):</b>\n\n"
                 f"{stat.str_refer()}\n"
