+++ pars_diary/handlers/base_commands.py
@@ -13,11 +13,11 @@ from aiogram.filters import Command
 from aiogram.types import Message
 from loguru import logger
 
-from utils.db import counter, get_cookie
-from utils.hw import hw
-from utils.keyboards import not_auth_keyboard
-from utils.messages import error, not_auth
-from utils.pars import Pars
+from pars_diary.utils.db import counter, get_cookie
+from pars_diary.utils.hw import hw
+from pars_diary.utils.keyboards import not_auth_keyboard
+from pars_diary.utils.messages import error, not_auth
+from pars_diary.utils.pars import Pars
 
 router = Router(name=__name__)
