+++ pars_diary/handlers/ch.py
@@ -5,9 +5,9 @@ from aiogram.filters import Command
 from aiogram.types import InlineKeyboardButton, InlineKeyboardMarkup, Message
 from loguru import logger
 
-from utils.ch import ch
-from utils.db import counter
-from utils.messages import error
+from pars_diary.utils.ch import ch
+from pars_diary.utils.db import counter
+from pars_diary.utils.messages import error
 
 router = Router(name=__name__)
 
@@ -32,10 +32,12 @@ async def ch_msg(msg: Message) -> None:
             answer[0],
             answer[1],
             "HTML",
-            reply_markup=InlineKeyboardMarkup(inline_keyboard=[
-                [InlineKeyboardButton(text="Подробнее", url=answer[2])],
-            ],
-        ))
+            reply_markup=InlineKeyboardMarkup(
+                inline_keyboard=[
+                    [InlineKeyboardButton(text="Подробнее", url=answer[2])],
+                ],
+            ),
+        )
 
     except Exception as e:
         # Отвечаем пользователю
