+++ pars_diary/handlers/chatgpt.py
@@ -7,9 +7,9 @@ from aiogram.filters import Command
 from aiogram.types import Message
 from loguru import logger
 
-from utils.ask_gpt import ask_gpt
-from utils.db import counter
-from utils.messages import error
+from pars_diary.utils.ask_gpt import ask_gpt
+from pars_diary.utils.db import counter
+from pars_diary.utils.messages import error
 
 router = Router(name=__name__)
