+++ pars_diary/models.py
@@ -1,12 +1,7 @@
 """Определение структур данных."""
 
-from __future__ import annotations
-
 from dataclasses import dataclass, field
-from typing import TYPE_CHECKING
-
-if TYPE_CHECKING:
-    from datetime import date
+from datetime import date
 
 
 @dataclass
