+++ pars_diary/utils/exceptions.py
