+++ pars_diary/utils/messages.py
