+++ pars_diary/utils/typing.py
